requirements.txt
Django==3.2
django-filter==23.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
pytz==2023.3
sqlparse==0.4.4
